Move culture names into its own file
