+++ Vic2ToHoI4Tests/Vic2ToHoI4Tests-Google.vcxproj.filters
@@ -917,6 +917,9 @@
     <ClCompile Include="..\Vic2ToHoI4\Source\HOI4World\Names\Names.cpp">
       <Filter>Vic2ToHoI4 files\HoI4\Names</Filter>
     </ClCompile>
+    <ClCompile Include="..\Vic2ToHoI4\Source\HOI4World\Names\CultureNames.cpp">
+      <Filter>Vic2ToHoI4 files\HoI4\Names</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <Filter Include="Vic2ToHoI4 files">
